Re-factored code and added simple styling
