+++ src/components/Game.scss
@@ -1,0 +1,6 @@
+.game {
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
+}
