+++ src/components/Task.scss
@@ -1,0 +1,24 @@
+.done-text {
+  color: white; 
+  background-color: black;
+}
+
+label {
+  font-size: 3em;
+}
+input[type=text] {
+  transition: all 0.30s ease-in-out;
+  outline: none;
+  padding: 3px 0px 3px 3px;
+  margin: 5px 1px 3px 0px;
+  border: 1px solid #DDDDDD;
+  font-size: 3em;
+  width: 100%;
+}
+
+input[type=text]:focus {
+  box-shadow: 0 0 5px rgba(81, 203, 238, 1);
+  padding: 3px 0px 3px 3px;
+  margin: 5px 1px 3px 0px;
+  border: 1px solid rgba(81, 203, 238, 1);
+}
